requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
discord.py==2.1.1
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isodate==0.6.1
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pafy==0.5.5
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.17
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
pytube==12.1.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.2
SecretStorage==3.3.1
six==1.16.0
streamlink==5.3.0
systemd-python==234
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.5.1
xdg==5
xkit==0.0.0
yarl==1.8.2
youtube-dl==2021.12.17
zipp==1.0.0
